refactor(dataviews): split facade modules
